Basic Physics Calculator
A calculator that calculates the basics physics quantities like speed,acceleration,velocity ,force etc. It also has the functionality that it can give the answer in  any unit we wanted to get.It also has the side navbar for the direct move to the page that has the parameter we wants to find.